feat!: replace to `httputil.ReverseProxy`
Closes #1054
