+++ .github/workflows/go.yml
@@ -14,7 +14,6 @@ jobs:
     strategy:
       matrix:
         go-version:
-          - '1.19'
           - '1.20'
           - '1.21'
