+++ lint-staged.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  '*.go': 'go fmt',
+  // '*.go': 'go fmt',
   '*.{js,jsx,ts,tsx}': 'eslint --fix',
   '*.{json,yml}': 'prettier --write'
 }
