+++ website/static/x/manael/internal/testutil.html
@@ -9,7 +9,7 @@
   name="go-source"
 />
 <meta
-  content="0; url=https://godoc.org/manael.org/x/manael/internal/testutil"
+  content="0; url=https://pkg.go.dev/manael.org/x/manael/internal/testutil"
   http-equiv="refresh"
 />
 <title>manael</title>
