+++ website/static/x/manael/v2/cmd/manael.html
@@ -9,7 +9,7 @@
   name="go-source"
 />
 <meta
-  content="0; url=https://pkg.go.dev/manael.org/x/manael/cmd/manael"
+  content="0; url=https://pkg.go.dev/manael.org/x/manael/v2/cmd/manael"
   http-equiv="refresh"
 />
 <title>manael</title>
